+++ go/runtime/primitives/prim_complex_test.go
@@ -175,3 +175,49 @@ func TestDenominator(t *testing.T) {
 		})
 	}
 }
+
+func TestMakeRectangularExtended(t *testing.T) {
+	tcs := []schemeCodeErrorTestCase{
+		{
+			name: "make-rectangular with integers",
+			code: `(make-rectangular 3 4)`,
+		},
+		{
+			name: "make-rectangular with floats",
+			code: `(make-rectangular 3.0 4.0)`,
+		},
+		{
+			name: "make-rectangular with rationals",
+			code: `(make-rectangular 1/2 3/4)`,
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			result, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+			qt.Assert(t, result, qt.IsNotNil)
+		})
+	}
+}
+
+func TestMagnitudeExtended(t *testing.T) {
+	tcs := []schemeCodeErrorTestCase{
+		{
+			name: "magnitude of rational",
+			code: `(magnitude 3/4)`,
+		},
+		{
+			name: "magnitude of negative",
+			code: `(magnitude -5)`,
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			result, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+			qt.Assert(t, result, qt.IsNotNil)
+		})
+	}
+}
