+++ go/runtime/primitives/prim_numeric_extra_test.go
@@ -508,3 +508,64 @@ func TestTruncateRemainder(t *testing.T) {
 		})
 	}
 }
+
+func TestRationalizeMoreCases(t *testing.T) {
+	tcs := []schemeCodeErrorTestCase{
+		{
+			name: "rationalize integer",
+			code: `(rationalize 5 0)`,
+		},
+		{
+			name: "rationalize float",
+			code: `(rationalize 3.14159 0.001)`,
+		},
+		{
+			name: "rationalize with larger tolerance",
+			code: `(rationalize 3.14159 0.5)`,
+		},
+		{
+			name: "rationalize negative",
+			code: `(rationalize -3.14159 0.001)`,
+		},
+		{
+			name: "rationalize zero",
+			code: `(rationalize 0 0.1)`,
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			_, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+		})
+	}
+}
+
+func TestExactInexactExtended(t *testing.T) {
+	tcs := []schemeCodeErrorTestCase{
+		{
+			name: "exact on rational",
+			code: `(exact 1/2)`,
+		},
+		{
+			name: "exact on integer",
+			code: `(exact 42)`,
+		},
+		{
+			name: "inexact on integer",
+			code: `(inexact 42)`,
+		},
+		{
+			name: "inexact on rational",
+			code: `(inexact 1/2)`,
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			result, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+			qt.Assert(t, result, qt.IsNotNil)
+		})
+	}
+}
