+++ go/runtime/primitives/prim_string_test.go
@@ -333,3 +333,42 @@ func TestStringToNumber(t *testing.T) {
 		})
 	}
 }
+
+func TestStringAppendExtended(t *testing.T) {
+	tcs := []struct {
+		name     string
+		code     string
+		expected string
+	}{
+		{
+			name:     "string-append five strings",
+			code:     `(string-append "a" "b" "c" "d" "e")`,
+			expected: "abcde",
+		},
+		{
+			name:     "string-append two strings",
+			code:     `(string-append "hello" "world")`,
+			expected: "helloworld",
+		},
+		{
+			name:     "string-append with space",
+			code:     `(string-append "hello" " " "world")`,
+			expected: "hello world",
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			result, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+			s, ok := result.(*values.String)
+			qt.Assert(t, ok, qt.IsTrue)
+			qt.Assert(t, s.Value, qt.Equals, tc.expected)
+		})
+	}
+}
+
+func TestStringAppendWithNonString(t *testing.T) {
+	_, err := runSchemeCode(t, `(string-append "hello" 42)`)
+	qt.Assert(t, err, qt.IsNotNil)
+}
