+++ go/runtime/primitives/prim_trig_test.go
@@ -342,3 +342,55 @@ func TestLog(t *testing.T) {
 		})
 	}
 }
+
+func TestTrigWithRationals(t *testing.T) {
+	tcs := []schemeCodeErrorTestCase{
+		{
+			name: "sin with rational",
+			code: `(sin 1/2)`,
+		},
+		{
+			name: "cos with rational",
+			code: `(cos 1/2)`,
+		},
+		{
+			name: "tan with rational",
+			code: `(tan 1/4)`,
+		},
+		{
+			name: "asin with rational",
+			code: `(asin 1/2)`,
+		},
+		{
+			name: "acos with rational",
+			code: `(acos 1/2)`,
+		},
+		{
+			name: "atan with rational",
+			code: `(atan 1/2)`,
+		},
+		{
+			name: "atan2 with rationals",
+			code: `(atan 1/2 3/4)`,
+		},
+		{
+			name: "log with rational",
+			code: `(log 1/2)`,
+		},
+		{
+			name: "log with rational base",
+			code: `(log 8 2)`,
+		},
+		{
+			name: "exp with rational",
+			code: `(exp 1/2)`,
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			_, err := runSchemeCode(t, tc.code)
+			qt.Assert(t, err, qt.IsNil)
+		})
+	}
+}
